Split users controller
